feat: Adicionando rota de exclusão de acessorioVeiculo e validação na exclusão que possui veículos e acessorios associados, excluir o relacionamento para não ficar registros inválidos no banco

+++ backend/api/Rotas/Veiculos/Acessorios/exclusaoAcessorios.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const deletarAcessorios = express();
 const tabelaAcessorio = require("../../../../Database/Tabelas/Acessorios/acessorios.js");
+const tabelaAcessorioVeiculo = require("../../../../Database/Tabelas/Acessorios_veiculos/acessoriosVeiculos.js");
 
 deletarAcessorios.delete(
   "/deletar-acessorio/:id",
@@ -13,21 +14,36 @@ deletarAcessorios.delete(
         },
       });
 
+      const acessorioVeiculo = await tabelaAcessorioVeiculo.findOne({
+        where: {
+          idAcessorio: id,
+        },
+      });
+
       if (!acessorio) {
         return response
           .status(400)
           .json({ Error: "Acessório não encontrado!!" });
       }
 
-      tabelaAcessorio.destroy({
-        where: {
-          idAcessorio: id,
-        },
-      });
+      //Verifica se possui acessórios associados a um veículo
+      if (acessorioVeiculo) {
+        tabelaAcessorioVeiculo.destroy({
+          where: {
+            idAcessorio: id,
+          },
+        });
 
-      return response
-        .status(200)
-        .json({ message: "Acessório excluído com sucesso!!" });
+        tabelaAcessorio.destroy({
+          where: {
+            idAcessorio: id,
+          },
+        });
+
+        return response
+          .status(200)
+          .json({ message: "Acessório excluído com sucesso!!" });
+      }
     } catch (error) {
       console.log(error);
       return response
